backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
alembic==1.13.2

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings==2.5.2
email-validator==2.2.0

# Scheduler
apscheduler==3.10.4

# Poker engine
pypokerengine==1.0.1
treys==0.1.8

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Linting
ruff==0.6.9
